day38/day38_cdc_debezium/requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
psycopg2-binary==2.9.9
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
avro-python3==1.10.2
faker==24.0.0
prometheus-client==0.20.0
pytest==8.1.1
pytest-asyncio==0.23.6
